Changed Order Get/Remove to take multiple ID's. Added Note that for remove ID's.
Changed Order Get/Remove to take multiple ID's. Added Note that for remove ID's
should start from the highest and go to the lowest otherwise unknown results
will occur.

Change some of the Order Descriptions. Lee could you check that this seems
okay? I've made them a bit more helpful on the client side (IE So I don't have
to keep asking the server if the value is valid.)

I've also added the "Relative Coordinates" (should it replace Absolute
coordinates which are just relative to the Universe?) and "Range".

If you give the go ahead I'll start implimenting it in py-netlib.